requirements.txt
absl-py==1.1.0
accelerate==0.19.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2023.5.7
charset-normalizer==2.1.0
click==8.1.3
cmake==3.26.3
cycler==0.11.0
easydict==1.9
filelock==3.7.1
fonttools==4.34.4
fsspec==2023.5.0
gdown==4.5.1
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
huggingface-hub==0.14.1
idna==3.3
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
lit==16.0.3
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.5.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
protobuf==3.19.4
psutil==5.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-crf==0.7.2
pytz==2022.1
PyYAML==6.0
regex==2023.5.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.13.3
tqdm==4.64.0
transformers==4.29.1
triton==2.0.0
typing_extensions==4.2.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.0
